@@ -1,7 +1,7 @@
 module.exports = {
   semi: true,
-  trailingComma: 'all',
+  trailingComma: "es5",
   singleQuote: true,
   printWidth: 120,
-  tabWidth: 2
-}
+  tabWidth: 2,
+};
